toggling
Added a quick start screen (not done)
Added a end screen placeholder

Made 2 functions where you swap the contents of the page so you dont have 3 different html files for the start screen, question screen and end screen.

Added if(confirm()){} to the going back to start and finishing the project, these are in place so you dont accidentally click back to start without actually wanting to go there.
This also allows for the back button to not just go back a question but also go back to the start with a confirmation.