Saving answers in array
I added a new array in which the answers are kept/stored.

Also changed the prepare render up a small bit,
I made it so now all the buttons have their individual input (needed to get the original answers into the answerArray) and the prepareRender doesnt have any array editing anymore.

Finally I started on a skip button which is not done yet (the styling isnt done altough the functionality works (pressumably)).