Matching system now works definitively
Changed up the matching system, it now works in the way that I want and I only added one of my own arrays for the answers, technically this could also be added into the data.js array, now I will either keep it, add a third array into data.js or find a way to add it into one of the two existing arrays.

I probably wont add it into the file since I would like to keep it untouched.

Besides that I was guiding Raven with his project for the rest of the day.