Arranging result functional.
Updated the arranging of the results.
Basically the sorting of answers did not work and I finally gave in of trying to sort them ascending instead of descending because it wasnt working.

The solution i found was to just prepend the items instead of appending.

Also I added a "zero_container" this is because any party with a score of 0 their score attribute will be "undefined" so they dont get sorted properly, I made a seperate container which they get put into and appended to the bottom, not the best solution but it works and the application is simple enough where it shouldn't be an issue.