Calculate result and dataDump functions updated
First the calculate result,
this is completely revamped, redid the whole way of calculating, it's much much shorter and better, it's also clearer to the reader.

Datadump is now shorter and simply better, i changed the way scores were added so now it doesnt need an exeption for parties with score 0.

Finally removed 1 party from the parties array in data.js because there were 2 of it. (libertarische partij)