added weight button with functionality + dynamic total points
Created a weight button to add 2 points instead of 1,
then created a function which dynamically changes the total amount of points used in the percentage calculation.

This is so that when you have 0 checks and answer everything pro then the top party will have 60% if you check everything double points and pro then the top party will still have 60%. 

Otherwise if you check nothing the number is really low and if you check everything the percentages will be really high.