Optimization with result screen
Optimized the result screen,
scores now are displayed in a nicer way names and percentages are separated.

Had an issue where the background was ignoring the results. This is fixed now.

Made it so that the render that worked on displaying the screen to show the buttons for party size to display in answer now is a separate function and now can be called more easily.

In the answer displaying I used a function, this now is placed in the function that first called it. This allows it to separate the percentage and name.